% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{uniscape_rstudio}
\alias{uniscape_rstudio}
\alias{escapeAllAddin}
\alias{escapeStringsAddin}
\alias{escapeFileAddin}
\alias{selectAll}
\title{RStudio addins provided by uniscape}
\usage{
escapeAllAddin()

escapeStringsAddin()

escapeFileAddin()

selectAll()
}
\description{
RStudio addins provided by uniscape
}
\section{Functions}{
\itemize{
\item \code{escapeAllAddin}: Escape all non-ASCII characters in selection

\item \code{escapeStringsAddin}: Escape non-ASCII characters inside R strings in selection

\item \code{escapeFileAddin}: Escape non-ASCII characters inside R strings in current source file

\item \code{selectAll}: Select all in current source file.
Internal use; not registered as RStudio addin.
}}

